refactor: #303
